remove bigint

+++ src/dlob/dlob_node.rs
@@ -1,6 +1,5 @@
 use crate::math::order::get_limit_price;
 use drift::state::{oracle::OraclePriceData, user::Order};
-use num_bigint::BigInt;
 use solana_sdk::pubkey::Pubkey;
 
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
@@ -20,7 +19,7 @@ pub(crate) enum SortDirection {
 }
 
 pub(crate) trait DLOBNode {
-    fn get_price(&self, oracle_price_data: OraclePriceData, slot: u64) -> BigInt;
+    fn get_price(&self, oracle_price_data: OraclePriceData, slot: u64) -> u64;
     fn is_vamm_node(&self) -> bool;
     fn is_base_filled(&self) -> bool;
     fn get_sort_value(&self, order: &Order) -> Option<i128>;
@@ -118,7 +117,7 @@ impl Node {
 }
 
 impl DLOBNode for Node {
-    fn get_price(&self, oracle_price_data: OraclePriceData, slot: u64) -> BigInt {
+    fn get_price(&self, oracle_price_data: OraclePriceData, slot: u64) -> u64 {
         match self {
             Node::OrderNode(order_node) => order_node.get_price(oracle_price_data, slot),
             Node::VAMMNode(vamm_node) => vamm_node.get_price(oracle_price_data, slot),
@@ -200,7 +199,7 @@ impl OrderNode {
 }
 
 impl DLOBNode for OrderNode {
-    fn get_price(&self, oracle_price_data: OraclePriceData, slot: u64) -> BigInt {
+    fn get_price(&self, oracle_price_data: OraclePriceData, slot: u64) -> u64 {
         get_limit_price(&self.order, &oracle_price_data, slot, None)
     }
 
@@ -257,8 +256,8 @@ impl VAMMNode {
 }
 
 impl DLOBNode for VAMMNode {
-    fn get_price(&self, _oracle_price_data: OraclePriceData, _slot: u64) -> BigInt {
-        BigInt::from(self.price)
+    fn get_price(&self, _oracle_price_data: OraclePriceData, _slot: u64) -> u64 {
+        self.price
     }
 
     fn is_vamm_node(&self) -> bool {
